Conflicts:
	cmake/QtCreatorIDEBranding.cmake
	qbs/modules/qtc/qtc.qbs
	qtcreator_ide_branding.pri
	src/plugins/mcusupport/mcusupportrunconfiguration.cpp
	src/plugins/python/pythonproject.cpp
	src/plugins/qmakeprojectmanager/qmakestep.cpp
	src/plugins/qmlprojectmanager/qmlproject.cpp
	src/plugins/qmlprojectmanager/qmlprojectrunconfiguration.cpp

+++ src/plugins/android/androidrunner.cpp
@@ -224,7 +224,7 @@ void AndroidRunner::remoteErrorOutput(const QString &output)
 }
 
 void AndroidRunner::handleRemoteProcessStarted(Utils::Port gdbServerPort,
-                                               const QUrl &qmlServer, int pid)
+                                               const QUrl &qmlServer, qint64 pid)
 {
     m_pid = ProcessHandle(pid);
     m_gdbServerPort = gdbServerPort;
